pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chest-xray-pneumonia-mlops"
version = "0.1.0"
description = "End-to-end MLOps system for chest X-ray pneumonia detection"
authors = [
    {name = "MLOps Team", email = "mlops@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Core ML libraries
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    
    # Image processing
    "opencv-python>=4.5.0",
    "Pillow>=8.0.0",
    "albumentations>=1.3.0",
    
    # MLOps tools
    "mlflow>=2.0.0",
    "dvc>=3.0.0",
    "optuna>=3.0.0",
    
    # API and web
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    
    # Database and storage
    "psycopg2-binary>=2.9.0",
    "boto3>=1.26.0",
    "minio>=7.1.0",
    
    # Monitoring and logging
    "prometheus-client>=0.16.0",
    "grafana-api>=1.0.0",
    
    # Utilities
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "tqdm>=4.64.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/example/chest-xray-pneumonia-mlops"
Repository = "https://github.com/example/chest-xray-pneumonia-mlops"
Issues = "https://github.com/example/chest-xray-pneumonia-mlops/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["data_pipeline*", "training*", "model_registry*", "deployment*", "monitoring*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=. --cov-report=html --cov-report=term-missing"